+++ lib/loggable_activity/data_owner.rb
@@ -1,15 +1,16 @@
 # frozen_string_literal: true
 
 require 'active_record'
-require 'awesome_print'
 
 module LoggableActivity
+  # This class represends an additional data owner for a record.
+  # For it to kick in, the data_owner configuration has to be set to true in the loggable_activity.yaml file.
   class DataOwner < ActiveRecord::Base
     self.table_name = 'loggable_data_owners'
-    # belongs_to :record, class_name: 'LoggableActivity::Record'
     belongs_to :record, polymorphic: true, optional: true
     belongs_to :encryption_key, class_name: '::LoggableActivity::EncryptionKey'
 
+    # When a record is deleted, all data owner added to the record is also deleted.
     def mark_as_deleted!
       encryption_key.mark_as_deleted!
     end