+++ lib/loggable_activity/encryption_key.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require 'active_record'
-require 'awesome_print'
+
 module LoggableActivity
   # This class represents the encryption key used to unlock the data for one payload.
   # When deleted, only the encryption_key field is deleted.
@@ -19,6 +19,7 @@ module LoggableActivity
       update(secret_key: nil)
     end
 
+    # check if the encryption key is deleted
     def deleted?
       secret_key.nil?
     end