+++ spec/factories/loggable_activities.rb
@@ -2,7 +2,7 @@
 
 # spec/factories/loggable_activities.rb
 FactoryBot.define do
-  factory :activity, class: ::LoggableActivity::Activity do
+  factory :activity, class: LoggableActivity::Activity do
     action { 'user.show' }
     actor factory: :user # Assuming you have a :user factory for the actor
     record factory: :user # Replace :some_record with an appropriate factory for the record